+++ src/docs/history.md
@@ -1,4 +1,13 @@
 
+1.8.6
+- hm3 syntax -> separate via '|syntax|'
+- fix for #79 (.csv and .svg files)
+- column names - updated correctly even on first load
+- show errors in status bar: if any
+- handle "Access denied" file error (show error in status bar, try again in 5 seconds)
+- handle the case when user is drag and dropping a Windows Event Log (show an error)
+
+
 1.8.5
 - fixed #78 Syntax not working with pipe symbol separators 
 - added "If line starts with tab, assume from previous line" log setting (for line-by-line files)
@@ -13,6 +22,7 @@
 
 1.8.2
 - small fixes
+
 
 1.8.1
 - to release to public